Make queue and priority queue iterable
A bit tricky with the priority queue, as the binary heap array is a
partially sorted implicit data structure and there isn’t a
straightforward way of traversing it without introducing more complexity
than is worth it here. So the current solution just slices off the

Note that it’s potentially possible to sort the heap itself, as the
shape property would be retained and the bubble down/up methods will
take care of any subsequent insertions. One issue with doing that in the
current state is that the root of the heap is at index [1] rather than
index [0], so there’s a blank/undefined sentinel value in that first
position—which provides some benefits in terms of simplifying arithmetic,
but has also led to more complexity in this implementation than would be
idea. May revisit this at some stage in the future, as there will be
performance and optimization benefits to changing the heap array
structure to be more predictable to the JavaScript engine.