Add poll() method which returns the priority as well as the element
Change to the API to introduce poll() and shift() which are intended to
replace pop(). This should provide greater clarity when using the API.
pop() makes it feel too much like a stack or array.

Priority elements are still stored as two element arrays, but the order
has swapped so that the reference is at [0] and the priority is at [1].
As a future change, it might be worthwhile to replace these arrays with
a struct-like object that provides a common shape for each element in
the list.