Add clear() method to queue and priority queue
It’s now possible to wipe all the elements in the collections with a
call to clear(). Still not really sure if it’s worth implementing remove
methods for individual elements. I haven’t seen a use for it so far when
working with these classes, but maybe it would make sense at some point.